'添加查看所有用户接口'

+++ pblog/src/main/java/com/pblogteam/pblog/vo/UserVO.java
@@ -25,6 +25,6 @@ public class UserVO
     private String photoUrl;
     private String description;
     private Integer attentionCount;
-    private Byte privilege;
+    private Integer privilege;
     private boolean myAttention;
 }
